requirements.txt
celery[redis]
django-celery-beat
django_apscheduler
erniebot-agent[all]
langchain~=0.2.7
django~=4.2.11
pypdf2~=3.0.1
pyyaml~=6.0.1
tqdm~=4.66.2
rsa~=4.9
colorama~=0.4.6
langchain_community
pypdf
spacy
faiss-cpu
langchain_core
asgiref~=3.7.2
apscheduler~=3.10.4